1,初步完成客户端代码重构 2 修复了一个启动bug 3 修复了两个内存泄漏问题 备注： 1 发现了一个多线程发送消息会产生冲突bug

+++ RemoteClient/ClientSocket.h
@@ -189,7 +189,7 @@ public:
 		while (true)
 		{
 			size_t len = recv(m_sock, buffer + index, BUFFER_SIZE - index, 0);
-			if ((len <= 0 ) && (index <=0)) {
+			if (((int)len <= 0 ) && ((int)index <=0)) {
 				return -1;
 			}
 			index += len;
